Added initial comments to some client managers
