+++ src/client/gui/bb_widget/activity_widget/activity.rs
@@ -9,7 +9,6 @@ use crate::client::gui::bb_widget::activity_widget::date_utils::{
 };
 use crate::client::gui::bb_widget::widget_utils::INDENT;
 use crate::client::gui::user_interface::Message;
-use crate::common::exercise_mod::exercise::Exercise;
 use crate::common::mascot_mod::mascot::Mascot;
 use crate::common::mascot_mod::mascot_trait::MascotTrait;
 use chrono::{Datelike, Duration, Local, NaiveDate};
@@ -346,17 +345,4 @@ where
     fn from(activity_widget: ActivityWidget) -> Self {
         Self::new(activity_widget)
     }
-}
-
-pub fn calculate_activity_data(exercise_data: &Vec<Exercise>) -> ActivityData {
-    let mut map: ActivityData = HashMap::new();
-
-    for exercise in exercise_data {
-        for (date, set) in &exercise.sets {
-            map.entry(*date)
-                .and_modify(|entry| *entry += set.len() as AmountOfSets)
-                .or_insert(set.len() as AmountOfSets);
-        }
-    }
-    map
 }
